Minor changes to allow running spanner integration tests locally
You'll need to update the specific test you want to run to use a CloudMySQLResourceManager instead of the default
MySQLResourceManager it's likely using. That resource manager can be used on the local side, and it also exposes a
method `forDataflow()` that can be used to create a resource manager for use on the dataflow side (which is what should
be passed to `launchDataflowJob()`).

You'll also need a running Cloud MySQL or PostgreSQL instance with both public and private connectivity enabled. When running the
test, pass in the following parameters:
    -DcloudProxyHost=<local cloud sql proxy host>
    -DcloudProxyPort=<local cloud sql proxy port>
    -DcloudProxyMySqlPort=<local cloud sql proxy port> # (only for MySQL)
    -DcloudProxyPostgresPort=<local cloud sql proxy port> # (only for PostgreSQL)
    -DcloudPrivateHost=<cloud sql private host>
    -DcloudPrivatePort=<cloud sql private port>
    -DcloudProxyUsername=<username> (can be omitted, defaults to root for MySQL and postgres for PostgreSQL)
    -DcloudProxyPassword=<password>